refactor(registry): 重构注册表存储结构以使用独立的存储模块
