+++ crates/puniyu_registry/src/command.rs
@@ -1,7 +1,8 @@
-use crate::store::STORE;
+mod store;
+use store::CommandStore;
 use puniyu_types::command::CommandBuilder;
-use std::sync::Arc;
-
+use std::sync::{Arc, LazyLock};
+static STORE: LazyLock<CommandStore> = LazyLock::new(CommandStore::new);
 #[derive(Clone)]
 pub struct Command {
 	/// 插件名称
@@ -34,34 +35,34 @@ impl CommandRegistry {
 			prefix: prefix.map(|s| s.to_string()),
 			builder,
 		};
-		STORE.command().insert(command);
+		STORE.insert(command);
 	}
 
 	pub fn remove_with_id(id: u64) {
-		STORE.command().remove_with_id(id);
+		STORE.remove_with_id(id);
 	}
 
 	pub fn remove_with_name(name: &str) {
-		STORE.command().remove_with_name(name)
+		STORE.remove_with_name(name)
 	}
 
 	pub fn remove_with_plugin_name(plugin_name: &str) {
-		STORE.command().remove_with_plugin_name(plugin_name)
+		STORE.remove_with_plugin_name(plugin_name)
 	}
 	pub fn get_with_id(id: u64) -> Option<Arc<Command>> {
-		STORE.command().get_with_id(id)
+		STORE.get_with_id(id)
 	}
 
 	pub fn get_with_name(name: &str) -> Option<Arc<Command>> {
-		STORE.command().get_with_name(name)
+		STORE.get_with_name(name)
 	}
 
 	pub fn get_with_plugin(plugin_name: &str, name: &str) -> Option<Arc<Command>> {
-		STORE.command().get_with_plugin(plugin_name, name)
+		STORE.get_with_plugin(plugin_name, name)
 	}
 
 	pub fn commands() -> Vec<Arc<Command>> {
-		STORE.command().all()
+		STORE.all()
 	}
