+++ crates/puniyu_registry/src/cooldown.rs
@@ -1,30 +1,33 @@
 mod types;
+mod store;
+use store::CooldownStore;
+
+use std::sync::LazyLock;
 use std::time::Duration;
 
 pub use types::CooldownScope;
 
-use crate::store::STORE;
-
+static STORE: LazyLock<CooldownStore> = LazyLock::new(CooldownStore::new);
 pub struct CooldownRegistry;
 
 impl CooldownRegistry {
 	/// 检查是否处于冷却期
 	pub fn is_cooling_down(scope: &CooldownScope) -> bool {
-		STORE.cooldown().is_cooling_down(scope)
+		STORE.is_cooling_down(scope)
 	}
 
 	/// 设置冷却记录
 	pub fn set_cooldown(scope: &CooldownScope, duration: Duration) {
-		STORE.cooldown().set_cooldown(scope, duration);
+		STORE.set_cooldown(scope, duration);
 	}
 
 	/// 清除冷却记录
 	pub fn clear_cooldown(scope: &CooldownScope) {
-		STORE.cooldown().clear_cooldown(scope);
+		STORE.clear_cooldown(scope);
 	}
 
 	/// 清理所有过期的冷却记录
 	pub fn cleanup_expired() {
-		STORE.cooldown().cleanup_expired();
+		STORE.cleanup_expired();
 	}
 }
