+++ crates/puniyu_registry/src/handler.rs
@@ -1,24 +1,27 @@
-use crate::store::STORE;
-use puniyu_types::handler::Handler;
-use std::sync::Arc;
+mod store;
+use store::HandlerStore;
 
+use puniyu_types::handler::Handler;
+use std::sync::{Arc, LazyLock};
+
+static STORE: LazyLock<HandlerStore> = LazyLock::new(|| HandlerStore::new());
 #[derive(Default)]
 pub struct HandlerRegistry;
 
 impl HandlerRegistry {
 	pub fn register(handler: Arc<dyn Handler>) {
-		STORE.handler().register(handler)
+		STORE.register(handler)
 	}
 
 	pub fn unregister(name: &str) -> bool {
-		STORE.handler().unregister(name)
+		STORE.unregister(name)
 	}
 
 	pub fn get(name: &str) -> Option<Arc<dyn Handler>> {
-		STORE.handler().get(name)
+		STORE.get(name)
 	}
 
 	pub fn handlers() -> Vec<Arc<dyn Handler>> {
-		STORE.handler().all()
+		STORE.all()
 	}
 }
