+++ crates/puniyu_registry/src/hook.rs
@@ -1,25 +1,28 @@
-use std::sync::Arc;
+mod store;
 use puniyu_types::hook::HookBuilder;
-use crate::store::STORE;
+use std::sync::{Arc, LazyLock};
+use store::HookStore;
+
+static STORE: LazyLock<HookStore> = LazyLock::new(HookStore::new);
 
 pub struct HookInfo {
-    pub index: u64,
-    pub builder: Arc<dyn HookBuilder>,
+	pub index: u64,
+	pub builder: Arc<dyn HookBuilder>,
 }
 
 pub struct HookRegistry;
 
 impl HookRegistry {
-    pub fn register(hook: Arc<dyn HookBuilder>) {
-        STORE.hook().insert(hook);
-    }
-    pub fn unregister(index: u64) {
-        STORE.hook().remove_with_index(index)
-    }
-    pub fn get(name: &str) -> Option<HookInfo> {
-        STORE.hook().get(name)
-    }
-    pub fn all() -> Vec<HookInfo> {
-        STORE.hook().all()
-    }
+	pub fn register(hook: Arc<dyn HookBuilder>) {
+		STORE.insert(hook);
+	}
+	pub fn unregister(index: u64) {
+		STORE.remove_with_index(index)
+	}
+	pub fn get(name: &str) -> Option<HookInfo> {
+		STORE.get(name)
+	}
+	pub fn all() -> Vec<HookInfo> {
+		STORE.all()
+	}
 }
