+++ crates/puniyu_registry/src/lib.rs
@@ -1,4 +1,3 @@
-mod store;
 
 #[cfg(feature = "bot")]
 pub mod bot;
