+++ crates/puniyu_registry/src/server.rs
@@ -1,23 +1,27 @@
-use crate::store::STORE;
+mod store;
+
+use std::sync::LazyLock;
+use store::ServerStore;
 use puniyu_types::server::ServerType;
 
+static STORE: LazyLock<ServerStore> = LazyLock::new(ServerStore::new);
 #[derive(Debug, Default)]
 pub struct ServerRegistry;
 
 impl ServerRegistry {
-	pub fn insert(name: impl Into<String>, server: ServerType) {
-		STORE.server().insert(name, server);
+	pub fn insert(name: impl Into<String>, server: impl Into<ServerType>) {
+		STORE.insert(name, server.into());
 	}
 
 	pub fn get(name: &str) -> Option<ServerType> {
-		STORE.server().get(name)
+		STORE.get(name)
 	}
 
 	pub fn get_all() -> Vec<ServerType> {
-		STORE.server().all()
+		STORE.all()
 	}
 
 	pub fn remove(name: &str) {
-		STORE.server().remove(name);
+		STORE.remove(name);
 	}
 }
